Fixed data type error (converting between ints and floats)

+++ lib/queue_source_byte_impl.cc
@@ -103,7 +103,7 @@ namespace gr {
         std::vector<char> buffer;
 
         float index;
-        int data_size;
+        float data_size;
 
         if(queue->pop(temp_vector)){
